requirements.txt
# Telegram Bot Framework
aiogram==3.4.1

# Video Downloader
yt-dlp==2024.1.19

# Async File Operations
aiofiles==23.2.1

# Testing
pytest==8.0.0
pytest-asyncio==0.23.3
hypothesis==6.98.3

# Environment Variables
python-dotenv==1.0.1
